Feature/fix camera y (#35)
* Flip Y for camera

* Fix camera & rendering & box2d positions

* Let levelreader read in coordinates as doubles

+++ game/level/LevelDomain.h
@@ -22,15 +22,15 @@ namespace levelDomain {
     };
 
     struct tile {
-        int x;
-        int y;
+        double x;
+        double y;
         std::string sprite;
         PlatformKind kind;
     };
 
     struct spawnPoint {
-        int x;
-        int y;
+        double x;
+        double y;
     };
 
     struct level {
